Move cmd related files to cmd pkg
